requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
--no-binary psycopg2

amqp==5.0.9
    # via kombu
attrs==22.2.0
    # via jsonschema
beautifulsoup4==4.6.0
    # via -r requirements.in
billiard==3.6.4.0
    # via celery
celery[redis]==5.2.3
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2022.12.7
    # via
    #   -r requirements.in
    #   elasticsearch
    #   requests
cffi==1.14.4
    # via cryptography
chardet==3.0.4
    # via requests
click==8.0.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==3.3.2
    # via
    #   -r requirements.in
    #   pyopenssl
    #   requests
directory-healthcheck==2.0.0
    # via -r requirements.in
dj-database-url==0.4.1
    # via -r requirements.in
django==2.2.28
    # via
    #   -r requirements.in
    #   directory-healthcheck
    #   django-celery-beat
    #   django-filter
    #   django-redis
    #   django-timezone-field
    #   djangorestframework
    #   drf-spectacular
    #   sigauth
django-celery-beat==2.2.1
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-extensions==1.6.7
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-health-check==3.8.0
    # via directory-healthcheck
django-pglocks==1.0.2
    # via -r requirements.in
django-redis==4.10.0
    # via -r requirements.in
django-timezone-field==4.1.2
    # via django-celery-beat
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   drf-spectacular
    #   sigauth
drf-spectacular==0.25.1
    # via -r requirements.in
elasticsearch==7.9.1
    # via
    #   -r requirements.in
    #   elasticsearch-dsl
elasticsearch-dsl==7.3.0
    # via -r requirements.in
gunicorn==19.5.0
    # via -r requirements.in
idna==2.6
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.17.3
    # via drf-spectacular
kombu==5.2.3
    # via celery
mohawk==0.3.4
    # via sigauth
pillow==9.3.0
    # via -r requirements.in
prompt-toolkit==3.0.19
    # via click-repl
psycopg2==2.8.6
    # via -r requirements.in
pycparser==2.18
    # via cffi
pyopenssl==18.0.0
    # via requests
pyrsistent==0.19.3
    # via jsonschema
python-crontab==2.5.1
    # via django-celery-beat
python-dateutil==2.7.3
    # via
    #   elasticsearch-dsl
    #   python-crontab
pytz==2021.3
    # via
    #   celery
    #   django
    #   django-timezone-field
pyyaml==6.0
    # via drf-spectacular
raven==6.10.0
    # via -r requirements.in
redis==3.5.3
    # via
    #   celery
    #   django-redis
requests[security]==2.25.1
    # via -r requirements.in
sigauth==4.3.0
    # via -r requirements.in
six==1.11.0
    # via
    #   click-repl
    #   cryptography
    #   django-extensions
    #   elasticsearch-dsl
    #   mohawk
    #   pyopenssl
    #   python-dateutil
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.6
    # via
    #   -r requirements.in
    #   elasticsearch
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
waitress==2.1.2
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
whitenoise==3.2.2
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
